Small fixes in contact details; extract nav button
