+++ src/screens/SSIContactDetailsScreen/index.tsx
@@ -1,16 +1,15 @@
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
-import React, {FC} from 'react';
-import {ScreenRoutesEnum, StackParamList} from '../../types';
-import {navigationRef} from '../../navigation/rootNavigation';
-import {ContactInformationView} from '../../components/views/ContactInformationView';
-import {NavigationButton} from './components/NavigationButton';
-import {ContactDetailsNavigationSection, Container, Divider} from '../../styles/components/screens/SSIContactDetailsScreen';
 import {IssuerStatus} from '@sphereon/ui-components.core';
+import React from 'react';
+import {NavigationButton} from '../../components/NavigationButton';
+import {ContactInformationView} from '../../components/views/ContactInformationView';
+import {ContactDetailsNavigationSection, Container, Divider} from '../../styles/components/screens/SSIContactDetailsScreen';
+import {ScreenRoutesEnum, StackParamList} from '../../types';
 
 type Props = NativeStackScreenProps<StackParamList, ScreenRoutesEnum.CONTACT_DETAILS>;
 
-const SSIContactDetailsScreen: FC<Props> = (props: Props): JSX.Element => {
-  const {contact} = props.route.params;
+const SSIContactDetailsScreen = ({route, navigation}: Props) => {
+  const {contact} = route.params;
   const contactDetails = [
     {
       id: 'Name',
@@ -48,7 +47,6 @@ const SSIContactDetailsScreen: FC<Props> = (props: Props): JSX.Element => {
       value: contact.branding?.contacts,
     },
   ];
-
   return (
     <Container>
       <ContactInformationView
@@ -60,12 +58,9 @@ const SSIContactDetailsScreen: FC<Props> = (props: Props): JSX.Element => {
         status={IssuerStatus.VERIFIED}
       />
       <ContactDetailsNavigationSection>
-        <NavigationButton
-          label="Identities"
-          onPress={() => props.navigation.navigate(ScreenRoutesEnum.CONTACT_IDENTITIES, {identities: contact.identities})}
-        />
+        <NavigationButton label="Identities" onPress={() => navigation.push(ScreenRoutesEnum.CONTACT_IDENTITIES, {identities: contact.identities})} />
         <Divider />
-        <NavigationButton label="Contact Activities" onPress={() => props.navigation.navigate(ScreenRoutesEnum.CONTACT_ACTIVITY, {contact})} />
+        <NavigationButton label="Contact Activities" onPress={() => navigation.push(ScreenRoutesEnum.CONTACT_ACTIVITY, {contact})} />
       </ContactDetailsNavigationSection>
     </Container>
   );
